Add disconnect to auth client api
